chore(Lines): added LineOfPeriod
