+++ packages/chart/src/components/Line.tsx
@@ -1,27 +1,19 @@
 import {
+  AnimatedProp,
+  Color,
   LinearGradient,
   Path,
   PathProps,
   SkiaDefaultProps,
+  SkPath,
   vec,
 } from '@shopify/react-native-skia';
-import {
-  interpolateColor,
-  SharedValue,
-  useDerivedValue,
-} from 'react-native-reanimated';
-
-import {IChartTransition, ISharedChartState} from '../types';
-import {IPeriodsWithPaths} from '../utils';
 
 interface IChartProps
   extends Omit<SkiaDefaultProps<PathProps, 'start' | 'end'>, 'path'> {
-  periodsWithPaths: SharedValue<IPeriodsWithPaths>;
-  index: number;
+  chartPath: SkPath;
   width: number;
-  transition: IChartTransition;
-  state: ISharedChartState;
-  lineColors?: string[] | undefined;
+  colors?: AnimatedProp<Color[]>;
 }
 
 export const COLORS = [
@@ -35,40 +27,7 @@ export const COLORS = [
   '#FF8A57',
 ];
 
-export const Line = ({
-  periodsWithPaths,
-  index,
-  width,
-  transition,
-  state,
-  ...rest
-}: IChartProps) => {
-  const chartPath = useDerivedValue(() => {
-    const {current, next} = state.value;
-    const start = periodsWithPaths.value[current]?.lines[index]?.path;
-    const end = periodsWithPaths.value[next]?.lines[index]?.path;
-
-    if (start === undefined || end === undefined) {
-      throw Error('start === undefined || end === undefined SHEEE');
-    }
-
-    return end.interpolate(start, transition.value)!;
-  });
-
-  const colors = useDerivedValue(() => {
-    const {current, next} = state.value;
-    const start = periodsWithPaths.value[current]?.lines[index]?.colors ?? [];
-    const end = periodsWithPaths.value[next]?.lines[index]?.colors ?? [];
-
-    return end.map((endColor, i) =>
-      interpolateColor(
-        transition.value,
-        [0, 1],
-        [start[i] ?? endColor, endColor],
-      ),
-    );
-  });
-
+export const Line = ({chartPath, width, colors, ...rest}: IChartProps) => {
   return (
     <Path
       path={chartPath}
@@ -76,7 +35,11 @@ export const Line = ({
       strokeJoin="round"
       strokeWidth={2}
       {...rest}>
-      <LinearGradient start={vec(0, 0)} end={vec(width, 0)} colors={colors} />
+      <LinearGradient
+        start={vec(0, 0)}
+        end={vec(width, 0)}
+        colors={colors ?? COLORS}
+      />
     </Path>
   );
 };
